+++ src/scripts/pal5/data/1-query_data.py
@@ -35,7 +35,7 @@ except NameError:
 
 SAVE_LOC = paths.data / "pal5" / "gaia_ps1_xm_polygons.asdf"
 
-if snkmk["load_from_static"]:
+if snkmk["load_from_static"] and (paths.static / "pal5").exists():
     shutil.copyfile(paths.static / "pal5" / "gaia_ps1_xm_polygons.asdf", SAVE_LOC)
     sys.exit(0)
